"add new page"

+++ src/main/resources/static/bhande-billing-system/src/routes/User.js
@@ -1,16 +1,35 @@
-import React, { useEffect, useRef } from 'react'
+import React, {useEffect, useRef, useState} from 'react'
 import '../styles/home.scss';
-import { useSelector } from "react-redux";
+import {useDispatch, useSelector} from "react-redux";
 import { Button } from 'antd';
 import { deleteUser, editUser } from '../auth';
 import { useNavigate } from 'react-router-dom';
 import UserCard from '../components/usercomponents/UserCard';
 import Login from './Login';
+import ReactDOM from "react-dom";
+import Popup from "../components/Popup";
+import {actions} from "../state-stuff/reducer";
 
 const User = () => {
   const userData = useSelector(state => {
     return state.user.userData;
   });
+
+  const dispatch = useDispatch();
+
+  const popupRef = useRef(null);
+
+  const closePopup = ()=>{
+    if (popupRef.current) {
+      popupRef.current.style.display = "none"; // Hide popup
+
+      ReactDOM.createPortal (Login,popupRef.current);
+
+      console.log(popupRef.current);
+    } else {
+      console.error("popupRef is null!");
+    }
+  }
 
   const navigate = useNavigate();
   var columns = [];
@@ -22,33 +41,36 @@ const User = () => {
   }
 
   const deleteUserByUuid = (uuid) => {
-    const resp = deleteUser(uuid);
-    resp.then((data) => {
-      const { success } = data;
-      if (success) {
-        navigate("/user");
-        alert("Your Data has been deleted");
-      }
-
-      else alert("Something Happend")
-    })
+    const Resp=prompt("Hey Do you Really Wanted to Delete this User Then Type 'YES'");
+    if (Resp === "YES") {
+      const resp = deleteUser(uuid);
+      resp.then((data) => {
+        const {success} = data;
+        if (success) {
+          navigate("/user");
+          alert("Your Data has been deleted");
+        } else alert("Something Happend")
+      })
+    }
   }
 
   const eiditUser = (uuid)=>{
+    dispatch({type : actions.selectUser, payload : {uuid}});
 
-    console.log(uuid);
-    const resp = editUser(uuid);
-    resp.then((data)=>{
-      
-    })
+    popupRef.current.style.display = "flex";
 
   }
+  const [refresh,setRefresh] = useState(true);
+
+  useEffect(() => {
+
+  }, [refresh]);
 
   return (
-    <div className='home'>
-      <div>
-        {<table>
-          <thead>
+      <div className='home'>
+        <div>
+          {<table>
+            <thead>
             <tr>
               {
                 columns.map((col) => {
@@ -58,34 +80,36 @@ const User = () => {
               <th>Actions</th>
 
             </tr>
-          </thead>
-          <tbody>
+            </thead>
+            <tbody>
 
             {
               userData.map((data) => {
                 return (
-                  <tr>
-                    {
-                      columns.map((col) => {
-                        return <td>{data[col]}</td>;
-                      })
-                    }
-                    <th><Button onClick={() => {
-                      deleteUserByUuid(data.uuid)
-                    }}>Delete</Button> <Button
-                    
-                    onClick={()=>{
-                      eiditUser(data.uuid);
-                    }}>Edit</Button></th>
-                  </tr>)
+                    <tr>
+                      {
+                        columns.map((col) => {
+                          return <td>{data[col]}</td>;
+                        })
+                      }
+                      <th><Button onClick={() => {
+                        deleteUserByUuid(data.uuid)
+                      }}>Delete</Button> <Button
+
+                          onClick={() => {
+                            eiditUser(data.uuid);
+                          }}>Edit</Button></th>
+                    </tr>)
               })
             }
 
-          </tbody>
-        </table>
-        }
+            </tbody>
+          </table>
+          }
+        </div>
+        <Popup popup = {popupRef} setRefresh = {setRefresh} refresh={refresh}/>
       </div>
-    </div>
+
   )
 }
